requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
ruff
torch
rembg
onnxruntime
pillow
requests